agents/requirements.txt
google-generativeai>=0.7.0
mysql-connector-python>=8.0.0
python-dotenv>=1.0.0
requests>=2.31.0